Fix decoder module errors and format specifiers for reliable builds
Remove `override` keyword from destructors in `rtl433_decoder` and `dsdfme_decoder` modules and correct format specifiers in `spectran_http_client` for uint64_t and size_t types.

+++ source_modules/spectran_http_source/src/spectran_http_client.cpp
@@ -50,10 +50,11 @@ void SpectranHTTPClient::setCenterFrequency(uint64_t freq) {
     // Make request
     net::http::RequestHeader rqhdr(net::http::METHOD_PUT, "/control", host);
     char buf[1024];
-    sprintf(buf, "{\"frequencyCenter\":%d,\"frequencySpan\":%d,\"type\":\"capture\"}", freq, _samplerate);
+    sprintf(buf, "{\"frequencyCenter\":%llu,\"frequencySpan\":%llu,\"type\":\"capture\"}",
+            (unsigned long long)freq, (unsigned long long)_samplerate);
     std::string data = buf;
     char lenBuf[16];
-    sprintf(lenBuf, "%d", data.size());
+    sprintf(lenBuf, "%llu", (unsigned long long)data.size());
     rqhdr.setField("Content-Length", lenBuf);
     controlHttp.sendRequestHeader(rqhdr);
     controlSock->sendstr(data);
